Fixed bug with role transfer
If the user.roleId is set to one of the built-in roles (regardless of
the value of user.role) then the user has that built-in role, not a
custom role.  If the user.roleId is not set to a built-in role then the
user has a custom role and we have to get the role from the role
manager.

It looks like user.role is (or can be) ignored.